backend/requirements.txt
flask
flask-cors
pandas
google-generativeai
transformers
torch
werkzeug
